Add css for component SelectPropertyEditor
