+++ frontend/packages/ux-editor/src/components/config/SelectPropertyEditor/SelectPropertyEditor.module.css
@@ -1,0 +1,27 @@
+.container {
+  display: grid;
+  gap: var(--fds-spacing-2);
+  padding-bottom: var(--fds-spacing-1);
+  align-items: center;
+}
+
+.viewMode {
+  margin-top: var(--fds-spacing-1);
+  margin-bottom: var(--fds-spacing-1);
+}
+
+.editSelectProperty {
+  display: flex;
+  flex-direction: row;
+  align-items: end;
+  gap: var(--fds-spacing-4);
+  margin-bottom: var(--fds-spacing-3);
+}
+
+.selectProperty {
+  flex: 1;
+}
+
+.viewSelectProperty {
+  padding-left: 0;
+}
